Fix: flatten src/src structure to src/
