--- develop_code/src/main/java/com/example/community_cr/CommunityCrApplication.java
@@ -1,13 +1,0 @@
-package com.example.community_cr;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CommunityCrApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CommunityCrApplication.class, args);
-	}
-
-}
