--- develop_code/src/main/java/com/example/community_cr/comment/controller/dto/request/UpdateCommentRequest.java
@@ -1,18 +1,0 @@
-package com.example.community_cr.comment.controller.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@Builder
-@AllArgsConstructor
-public class UpdateCommentRequest {
-	@NotBlank(message = "내용을 입력해주세요.")
-	@Size(max = 50)
-	private String content;
-}
