--- develop_code/src/main/java/com/example/community_cr/comment/entity/Comment.java
@@ -1,48 +1,0 @@
-package com.example.community_cr.comment.entity;
-
-import java.time.LocalDateTime;
-
-import com.example.community_cr.community.entity.Post;
-import com.example.community_cr.user.entity.User;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Builder
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
-
-	private String content;
-
-	@Builder.Default
-	private LocalDateTime createdAt = LocalDateTime.now();
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "post_id")
-	private Post post;
-
-	public void updateContent(String content) {
-		this.content = content;
-	}
-
-}
