--- develop_code/src/main/java/com/example/community_cr/common/config/WebConfig.java
@@ -1,38 +1,0 @@
-package com.example.community_cr.common.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-
-import io.micrometer.common.lang.NonNull;
-
-@Configuration
-public class WebConfig {
-
-	@Bean
-	public WebMvcConfigurer corsConfigurer() {
-		return new WebMvcConfigurer() {
-			@Override
-			public void addCorsMappings(@NonNull CorsRegistry registry) {
-				registry.addMapping("/**")
-					.allowedOrigins("*")
-					.allowedMethods("*")  // 모든 HTTP 메서드 허용
-					.allowedHeaders("*")
-					.allowCredentials(false)
-					.maxAge(3600L)
-					.exposedHeaders("Set-Cookie", "Authorization");
-			}
-		};
-	}
-
-	@Bean
-	public ObjectMapper objectMapper() {
-		ObjectMapper objectMapper = new ObjectMapper();
-		objectMapper.registerModule(new JavaTimeModule());
-		return objectMapper;
-	}
-}
