--- develop_code/src/main/java/com/example/community_cr/common/exception/GlobalExceptionHandler.java
@@ -1,32 +1,0 @@
-package com.example.community_cr.common.exception;
-
-import java.util.List;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.ObjectError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-	@ResponseStatus(HttpStatus.BAD_REQUEST)
-	@ExceptionHandler(MethodArgumentNotValidException.class)
-	public ResponseEntity<List<String>> handleMethodArgumentNotValidException(
-		MethodArgumentNotValidException exception) {
-		List<String> messages = exception.getBindingResult().getAllErrors().stream()
-			.map(ObjectError::getDefaultMessage)
-			.toList();
-		return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-			.body(messages);
-	}
-
-	@ResponseStatus(HttpStatus.BAD_REQUEST)
-	@ExceptionHandler(IllegalArgumentException.class)
-	public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException exception) {
-		return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-			.body(exception.getMessage());
-	}
-}
