--- develop_code/src/main/java/com/example/community_cr/like/repository/LikeService.java
@@ -1,7 +1,0 @@
-package com.example.community_cr.like.repository;
-
-public interface LikeService {
-	void likePost(long senderId, long postId);
-
-	void unlikePost(long senderId, long postId);
-}
