--- develop_code/src/test/java/com/example/community_cr/CommunityCrApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.community_cr;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CommunityCrApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
