+++ src/main/java/com/example/community_cr/community/controller/CommunityController.java
@@ -3,8 +3,11 @@ package com.example.community_cr.community.controller;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
@@ -15,6 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.example.community_cr.community.controller.dto.request.PostRequest;
+import com.example.community_cr.community.controller.dto.request.UpdatePostRequest;
 import com.example.community_cr.community.controller.dto.response.PostDetailResponse;
 import com.example.community_cr.community.controller.dto.response.PostResponse;
 import com.example.community_cr.community.service.CommunityService;
@@ -39,6 +43,34 @@ public class CommunityController {
 			postDetailResponse.orElseThrow(IllegalArgumentException::new));
 	}
 
+	@PatchMapping("/update/{postId}")
+	public ResponseEntity<PostDetailResponse> updatePost(
+		@RequestHeader("user-id") long userId,
+		@PathVariable long postId,
+		@RequestPart(name = "request", required = false) @Valid final Optional<UpdatePostRequest> updatePostRequest,
+		@RequestPart(value = "image", required = false) final Optional<MultipartFile> image) {
+		Optional<PostDetailResponse> postDetailResponse = Optional.empty();
+		if (updatePostRequest.isPresent() && image.isPresent()) {
+			postDetailResponse = communityService.updateCommunityPost(userId, postId, updatePostRequest.get(),
+				image.get());
+		} else if (updatePostRequest.isPresent()) {
+			postDetailResponse = communityService.updateCommunityPost(userId, postId, updatePostRequest.get());
+		} else if (image.isPresent()) {
+			postDetailResponse = communityService.updateCommunityPost(userId, postId, image.get());
+		}
+		return ResponseEntity.ok(
+			postDetailResponse.orElseThrow(IllegalArgumentException::new));
+	}
+
+	@DeleteMapping("/delete/{postId}")
+	public ResponseEntity<Void> deletePost(
+		@RequestHeader("user-id") long userId,
+		@PathVariable long postId) {
+		communityService.deletePost(userId, postId);
+		return ResponseEntity.status(HttpStatus.NO_CONTENT)
+			.build();
+	}
+
 	@GetMapping("/getPosts")
 	public ResponseEntity<List<PostResponse>> getPosts(
 		@RequestParam(name = "cursor", required = false, defaultValue = "0") long cursor,
@@ -47,9 +79,11 @@ public class CommunityController {
 			communityService.findAllCommunityPosts(cursor, count));
 	}
 
-	@GetMapping("/getPost/{id}")
-	public ResponseEntity<PostDetailResponse> getPost(@PathVariable long id) {
-		Optional<PostDetailResponse> postDetailResponse = communityService.findCommunityPostById(id);
+	@GetMapping("/getPost/{postId}")
+	public ResponseEntity<PostDetailResponse> getPost(
+		@RequestHeader("user-id") long userId,
+		@PathVariable long postId) {
+		Optional<PostDetailResponse> postDetailResponse = communityService.findCommunityPostById(userId, postId);
 		return ResponseEntity.ok(
 			postDetailResponse.orElseThrow(IllegalArgumentException::new));
 	}
