+++ src/main/java/com/example/community_cr/image/service/ImageService.java
@@ -5,5 +5,7 @@ import org.springframework.web.multipart.MultipartFile;
 public interface ImageService {
 	String upload(MultipartFile file);
 
+	void delete(String fileName);
+
 	String generatePresignedUrl(String fileName);
 }
