+++ src/main/java/com/example/community_cr/image/service/ImageServiceImpl.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.amazonaws.SdkClientException;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.GeneratePresignedUrlRequest;
 import com.amazonaws.services.s3.model.ObjectMetadata;
@@ -40,6 +41,15 @@ public class ImageServiceImpl implements ImageService {
 			s3Client.putObject(bucket, fileName, file.getInputStream(), metadata);
 			return fileName;
 		} catch (IOException e) {
+			throw new IllegalStateException("[Error] AWS S3 서비스 접근에 실패했습니다.");
+		}
+	}
+
+	@Override
+	public void delete(String fileName) {
+		try {
+			s3Client.deleteObject(bucket, fileName);
+		} catch (SdkClientException e) {
 			throw new IllegalStateException("[Error] AWS S3 서비스 접근에 실패했습니다.");
 		}
 	}
