+++ src/main/java/com/example/community_cr/like/entity/HeartId.java
@@ -1,5 +1,8 @@
 package com.example.community_cr.like.entity;
 
+import com.example.community_cr.community.entity.Post;
+import com.example.community_cr.user.entity.User;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -12,6 +15,20 @@ import lombok.NoArgsConstructor;
 public class HeartId {
 	private long user;
 	private long post;
+
+	public static HeartId of(Post post, User user) {
+		return HeartId.builder()
+			.post(post.getId())
+			.user(user.getId())
+			.build();
+	}
+
+	public static HeartId of(long postId, long userId) {
+		return HeartId.builder()
+			.post(postId)
+			.user(userId)
+			.build();
+	}
 
 	@Override
 	public boolean equals(Object object) {
